asdf

+++ ProjectRoot/bash/calibration.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
 
+CLEAN_SLATE="true"
+
+if ["$CLEAN_SLATE" = "true"]; then
+    rm -rf $HOME/OTA-C/ProjectRoot/storage/calibration/*
+    sleep 1
+fi
+
 cent_node="32B1728"
 leaf_node_serials=("32B172B" "32C793E" "32B1708" "32C7981" "337D42D" "32C79F7" "32C7920" "32C79BE" "32C79C6")
+# leaf_node_serials=("32B1708" "32C79F7" "32C7920" "32C79BE" "32C79C6")
 
 remote_nodes=("rpi4m1@192.168.5.241" "rpi4m2@192.168.5.242" "rpi4m3@192.168.5.243" "rpi4m4@192.168.5.244" "rpi4m4@192.168.5.244" "rpi4m5@192.168.5.245" "rpi4m5@192.168.5.245" "rpi4compute@192.168.5.246" "nuc01@192.168.5.248")
+# remote_nodes=("rpi4m3@192.168.5.243" "rpi4m5@192.168.5.245" "rpi4m5@192.168.5.245" "rpi4compute@192.168.5.246" "nuc01@192.168.5.248")
 
 timeout=900  # Set your timeout in seconds
 
@@ -18,8 +27,10 @@ do
     # Start two commands in detached screen sessions
     ssh ${node_name} "bash -c 'pkill -9 CA_calib'"
     sleep 1
-    ssh ${node_name} "bash -c 'rm -rf \$HOME/OTA-C/ProjectRoot/storage/calibration/*'"
-    sleep 1
+    if ["$CLEAN_SLATE" = "true"]; then
+        ssh ${node_name} "bash -c 'rm -rf \$HOME/OTA-C/ProjectRoot/storage/calibration/*'"
+        sleep 1
+    fi
     screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_leaf_${node_serial}.log" -S session_leaf_${node_serial} bash -c "ssh ${node_name} 'cd \$HOME/OTA-C/ProjectRoot/build && ./CA_calib leaf ${node_serial} ${cent_node}'"
     sleep 1
     screen -dmL -Logfile "$HOME/OTA-C/ProjectRoot/storage/calibration/session_cent_${node_serial}.log" -S session_cent_${node_serial} bash -c "cd $HOME/OTA-C/ProjectRoot/build/ && ./CA_calib cent ${cent_node} ${node_serial}"
